+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 1. [](#new)
     * Added `FileInterface::touch()` method
+    * Added optional `$field` parameter to `FlexMediaTrait::deleteMediaFile()` method
 2. [](#improved)
     * By default, add media to only pages which have been initialized in pages loop
 3. [](#bugfix)